Make sure that the (autofocus) plugins are found
Micro-Manager has this idea that "mmplugins/" and "mmautofocus/" are in
the current directory. This is not necessarily correct when it runs inside
Fiji, as it is convenient to call Fiji in the PATH and open a couple of
images in the current directory (which is usually outside of Fiji.app/ or
OpenSPIM.app/).

However, Micro-Manager offers the option to configure where the plugins
live, via setting a property. That is exactly what we do from now on.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>